+++ common/src/main/java/de/cech12/woodenhopper/client/WoodenHopperScreen.java
@@ -4,7 +4,7 @@ import de.cech12.woodenhopper.Constants;
 import de.cech12.woodenhopper.inventory.WoodenHopperContainer;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
-import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.RenderPipelines;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
@@ -31,6 +31,6 @@ public class WoodenHopperScreen extends AbstractContainerScreen<WoodenHopperCont
     protected void renderBg(@NotNull GuiGraphics guiGraphics, float partialTicks, int x, int y) {
         int i = (this.width - this.imageWidth) / 2;
         int j = (this.height - this.imageHeight) / 2;
-        guiGraphics.blit(RenderType::guiTextured, HOPPER_GUI_TEXTURE, i, j, 0, 0, this.imageWidth, this.imageHeight, 256, 256);
+        guiGraphics.blit(RenderPipelines.GUI_TEXTURED, HOPPER_GUI_TEXTURE, i, j, 0, 0, this.imageWidth, this.imageHeight, 256, 256);
     }
 }