+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [this versioning scheme](https://gist.github.com/cech12/69319028e88c50349a6b044000a6607b).
 
+## [1.21.6-3.6.0.0] - 2025-06-25
+### Changed
+- updated to Minecraft 1.21.6 (Fabric 0.127.1+1.21.6, NeoForge 21.6.16-beta, Forge 56.0.3)
+- updated Cloth Config support (19.0.147) (Fabric/Quilt)
+- updated ModMenu support (15.0.0-beta.2) (Fabric/Quilt)
+
 ## [1.21.5-3.5.0.0] - 2025-04-28
 ### Changed
 - Updated to Minecraft 1.21.5 (Fabric 0.120.0+1.21.5, NeoForge 21.5.47-beta, Forge 55.0.9)